requirements.txt
asgiref==3.8.1
attrs==25.3.0
autobahn==24.4.2
Automat==25.4.16
build==1.2.2.post1
certifi==2025.4.26
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
click==8.2.1
cloudinary==1.44.0
colorama==0.4.6
constantly==23.10.4
cryptography==44.0.3
daphne==4.1.2
dj-database-url==2.3.0
Django==5.2.1
django-cloudinary-storage==0.3.0
django-cors-headers==4.7.0
django-livereload-server==0.5.1
djangorestframework==3.16.0
hyperlink==21.0.0
idna==3.10
incremental==24.7.2
livereload==2.7.1
packaging==25.0
pillow==11.2.1
pip-tools==7.4.1
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pyOpenSSL==25.0.0
pyproject_hooks==1.2.0
requests==2.32.3
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
tornado==6.4.2
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
wheel==0.45.1
whitenoise==6.9.0
zope.interface==7.2

